voice-learning-module/ml-backend/requirements.txt
# ML Backend Requirements - Python 3.14 Compatible Minimal
# Flask API runs with these. Audio/ML features can be added later when packages support Python 3.14

# API Server
flask>=2.3.0
flask-cors>=4.0.0

# Basic Math/ML
numpy>=1.26.0
scipy>=1.10.0

# Config
python-dotenv>=0.21.0

# Utilities
requests>=2.31.0

# Development & Testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.7b0
flake8>=3.9.0
mypy>=0.910

# Jupyter for research
jupyter>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
